requirements.txt
streamlit
python-dotenv
openai
apify-client
psycopg2-binary
requests
